move wfs helpers in out of tools
